* origin/USDT买充优化:
  chore: 关闭新客服入口（非本分支内容）
  v: 版本更新
  feat: 充提指南选中效果
  feat: 【个人中心】买充提卖图标替换；【充提指南】自动播放改成滚动+高亮；【充值提现】底部banner进充提指南
  feat: 提现未办卡流程同H5；绑定手机去掉+86
  feat: 进游戏增加webApp标识
  update
  update
  feat: 更新UI
  feat: update
  feat: 接入小金库的H5支付；fix: 充币图标不显示
  feat: 【提币】增加充提指南入口；【首页】买充提卖增加图标
  feat: 充币页面优化（差小金库H5支付），买充提卖0星触发充提指南

# Conflicts:
#	HYNewNest/Features(主代码)/BASE/Controller/HYTabBarViewController.m
#	HYNewNest/Features(主代码)/HomeComponent/Controller/GameStartPlayViewController.m
#	HYNewNest/Features(主代码)/HomeComponent/View/CNUserInfoLoginView.m
#	HYNewNest/Features(主代码)/MeComponent/Controller/CNMineVC.xib

+++ HYNewNest/Features(主代码)/BASE/Networking/HYNetworkConfigManager.m
@@ -135,7 +135,7 @@
     [kKeywindow jk_makeToast:[IVHttpManager shareManager].gateway
                     duration:4
                     position:JKToastPositionCenter
-                       title:[NSString stringWithFormat:@"😄当前是%ld --【%@】",environment ,envName]];
+                       title:[NSString stringWithFormat:@"😄当前是%ld --【%@】",(long)environment ,envName]];
 #endif
 }
